WC: Separate API library from WooCommerce plugin For testing purposes, the API library was included in the WooCommerce plugin, which is highly unelegant for a production release. It is now in it's right place in src/php/ + The release target of the Makefile now includes the library in the zip.
